backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
pymongo==4.7.2
python-dotenv==1.0.1
certifi==2024.12.14
flask-compress==1.15
pandas==2.2.2
openai==1.30.1
google-cloud-vision==3.13.0
numpy==1.26.4
scikit-learn==1.5.0
gunicorn==21.2.0
waitress==2.1.2
Werkzeug==3.0.3
